+++ contracts/PriceFeedArb.sol
@@ -55,7 +55,7 @@ contract PriceFeedArb is Ownable, BaseMath {
     struct Response {
         int256 answer;
         uint64 blockNumber;
-        uint80 roundId;
+        uint256 roundId;
     }
 
     enum Status {
@@ -617,11 +617,11 @@ contract PriceFeedArb is Ownable, BaseMath {
     {
         // Try to get latest price data:
         try mainOracle.latestRoundData() returns (
-            uint80 roundId,
+            uint256 roundId,
             int256 answer,
             uint256 blockNumber,
             uint256 currentChainBN,
-            uint80 __roundId
+            uint256 __roundId
         ) {
             response.roundId = roundId;
             response.answer = answer;
@@ -641,11 +641,11 @@ contract PriceFeedArb is Ownable, BaseMath {
     {
         // Try to get latest price data:
         try backupOracle.latestRoundData() returns (
-            uint80 roundId,
+            uint256 roundId,
             int256 answer,
             uint256 blockNumber,
             uint256 currentChainBN,
-            uint80 __roundId
+            uint256 __roundId
         ) {
             response.roundId = roundId;
             response.answer = answer;
@@ -659,7 +659,7 @@ contract PriceFeedArb is Ownable, BaseMath {
     }
 
     function _getPrevOracleResponse(
-        uint80 _currentRoundId
+        uint256 _currentRoundId
     ) internal view returns (Response memory prevMainOracleResponse) {
         /*
          * NOTE: Oracle only offers a current decimals() value - there is no way to obtain the decimal precision used in a
@@ -668,11 +668,11 @@ contract PriceFeedArb is Ownable, BaseMath {
 
         // Try to get the price data from the previous round:
         try mainOracle.getRoundData(_currentRoundId - 1) returns (
-            uint80 roundId,
+            uint256 roundId,
             int256 answer,
             uint256 blockNumber,
             uint256 currentChainBN,
-            uint80 __roundId
+            uint256 __roundId
         ) {
             // If call to Hedgehog succeeds, return the response and success = true
             prevMainOracleResponse.roundId = roundId;
@@ -686,7 +686,7 @@ contract PriceFeedArb is Ownable, BaseMath {
     }
 
     function _getPrevBackupOracleResponse(
-        uint80 _currentRoundId
+        uint256 _currentRoundId
     ) internal view returns (Response memory prevBackupOracleResponse) {
         /*
          * NOTE: Oracle only offers a current decimals() value - there is no way to obtain the decimal precision used in a
@@ -695,11 +695,11 @@ contract PriceFeedArb is Ownable, BaseMath {
 
         // Try to get the price data from the previous round:
         try backupOracle.getRoundData(_currentRoundId - 1) returns (
-            uint80 roundId,
+            uint256 roundId,
             int256 answer,
             uint256 blockNumber,
             uint256 currentChainBN,
-            uint80 __roundId
+            uint256 __roundId
         ) {
             // If call to Hedgehog succeeds, return the response and success = true
             prevBackupOracleResponse.roundId = roundId;
